# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Company:
    properties:
      id:
        type: string
      location:
        $ref: '#/definitions/domain.CompanyLocation'
      logo:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.CompanySettings'
    type: object
  domain.CompanyLocation:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  domain.CompanySettings:
    properties:
      close_time:
        type: string
      currency:
        type: string
      measurement_unit:
        $ref: '#/definitions/domain.MeasurementUnit'
      open_time:
        type: string
    type: object
  domain.MeasurementUnit:
    enum:
    - km
    - mi
    type: string
    x-enum-varnames:
    - KilometresMeasurementUnit
    - MilesMeasurementUnit
  domain.Service:
    properties:
      company_id:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  http.companyCreateRequest:
    properties:
      location:
        $ref: '#/definitions/domain.CompanyLocation'
      logo:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.CompanySettings'
    type: object
  http.companyUpdateRequest:
    properties:
      location:
        $ref: '#/definitions/domain.CompanyLocation'
      logo:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.CompanySettings'
    type: object
  http.idResponse:
    properties:
      id:
        type: string
    type: object
  http.response:
    properties:
      message:
        type: string
    type: object
  http.serviceCreateRequest:
    properties:
      description:
        type: string
      duration:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  http.serviceUpdateRequest:
    properties:
      description:
        type: string
      duration:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
host: localhost:80
info:
  contact: {}
  title: Syncrepair API
  version: "1.0"
paths:
...
  /companies:
    delete:
      consumes:
      - application/json
      description: Delete company
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Delete
      tags:
      - companies
    get:
      consumes:
      - application/json
      description: Get company
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get
      tags:
      - companies
    post:
      consumes:
      - application/json
      description: Create new company
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.companyCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      summary: Create
      tags:
      - companies
    put:
      consumes:
      - application/json
      description: Update company
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.companyUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Update
      tags:
      - companies
  /services:
    get:
      consumes:
      - application/json
      description: Get all services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Service'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get all
      tags:
      - services
    post:
      consumes:
      - application/json
      description: Create service
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.serviceCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Create
      tags:
      - services
    put:
      consumes:
      - application/json
      description: Update service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.serviceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Update
      tags:
      - services
  /services/{id}:
    delete:
      consumes:
      - application/json
      description: Delete service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Delete
      tags:
      - services
    get:
      consumes:
      - application/json
      description: Get service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Service'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get by ID
      tags:
      - services
...
securityDefinitions:
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"